Remove team-switcher in favor of simple logout dropdown item
